Update Cypress snapshot?
Idk I think they might be flaky based on window size… might want to update the selector to be able to take a tighter screenshot without the black bars

I think the problem is that we don't actually have a convenient container element at the right position? All the images are doing their sizing themselves, I forget exactly why, I think so we can depend on the cover algorithm?